Travis: Allow to explicitly set XS branch to use
This gives us more flexibility, so we don't have to follow the name of
the main XS branch for each component, now we can explicitly set which
branch a given component branch belongs to. This can be done by adding
the GIT_BRANCH variable to the .travis.yml of the repo to be tested.

If it is not specified, then the name of the branch will be used, as
previously, this provides backward-compatibility, so that the builds of
the older branches will not get broken.

This is necessary to fix the Travis build of the 1.30-lcm branch of
xen-api.

Signed-off-by: Gabor Igloi <gabor.igloi@citrix.com>