[UPDATE] Make Travis happier
In the current state Travis CI is failing for several reasons. The default for Travis is currently to use Python3 rather than Python2. 'run.py' is still written for Python2.7 _(which is supposedly EOL since Jan 1st, 2020!)_. It probably need some love and a rewrite for Python3... Nevertheless now, after this commit, Travis gets to the bottom of the job, and just fails because instructed to clone a repo which is missing the Makefile.

Not sure if Travis is used _(I believe it is not!)_, but it should probably either make it work or removed the support entirely.